+++ lib/i18n/strings.g.dart
@@ -6,7 +6,7 @@
 /// Locales: 1
 /// Strings: 868
 ///
-/// Built on 2026-06-16 at 20:06 UTC
+/// Built on 2026-06-16 at 20:43 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import